refactor: move formbuilder/ to shared/formbuilder/
Part of #1235. Moves the RJSF-based form building system from
`src/formbuilder/` to `src/shared/formbuilder/` alongside the other
shared utilities. Updates all import paths; also fixes two relative
re-exports in index.ts to use @/src/ aliases for robustness.
