+++ packages/base/src/features/layers/forms/source/geotiffsource.tsx
@@ -3,13 +3,13 @@ import { showErrorMessage } from '@jupyterlab/apputils';
 import { RJSFSchema, UiSchema } from '@rjsf/utils';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
-import { SchemaForm } from '@/src/formbuilder/objectform/SchemaForm';
-import { FileSelectorWidget } from '@/src/formbuilder/objectform/fileselectorwidget';
+import { SchemaForm } from '@/src/shared/formbuilder/objectform/SchemaForm';
+import { FileSelectorWidget } from '@/src/shared/formbuilder/objectform/fileselectorwidget';
 import {
   processBaseSchema,
   removeFormEntry,
-} from '@/src/formbuilder/objectform/schemaUtils';
-import { useSchemaFormState } from '@/src/formbuilder/objectform/useSchemaFormState';
+} from '@/src/shared/formbuilder/objectform/schemaUtils';
+import { useSchemaFormState } from '@/src/shared/formbuilder/objectform/useSchemaFormState';
 import { deepCopy, getMimeType } from '@/src/tools';
 import type { ISourceFormProps } from './sourceform';
