--- packages/base/src/formbuilder/index.ts
@@ -1,4 +1,0 @@
-export * from './objectform/SchemaForm';
-export * from '../features/layers/forms/source';
-export * from '../features/layers/forms/layer';
-export * from './creationform';
