+++ packages/base/src/shared/formbuilder/objectform/components/OpacitySlider.tsx
